+++ netket/sampler/metropolis.py
@@ -28,7 +28,7 @@ from netket.hilbert import AbstractHilbert, ContinuousHilbert
 from netket.utils import mpi, wrap_afun
 from netket.utils.types import PyTree, DType
 
-from netket.utils.deprecation import deprecated, warn_deprecation
+from netket.utils.deprecation import warn_deprecation
 from netket.utils import struct
 
 from netket.utils.config_flags import config
@@ -526,32 +526,6 @@ class MetropolisSampler(Sampler):
             + f"machine_power = {sampler.machine_pow}, "
             + f"dtype = {sampler.dtype})"
         )
-
-
-@deprecated(
-    "The module function `sample_next` is deprecated in favor of the class method `sample_next`."
-)
-def sample_next(
-    sampler: MetropolisSampler,
-    machine: Union[Callable, nn.Module],
-    parameters: PyTree,
-    state: Optional[SamplerState] = None,
-) -> tuple[SamplerState, jnp.ndarray]:
-    """
-    Samples the next state in the Markov chain.
-
-    Args:
-        sampler: The Metropolis sampler.
-        machine: A Flax module or callable with the forward pass of the log-pdf.
-            If it is a callable, it should have the signature :code:`f(parameters, σ) -> jnp.ndarray`.
-        parameters: The PyTree of parameters of the model.
-        state: The current state of the sampler. If not specified, then initialize and reset it.
-
-    Returns:
-        state: The new state of the sampler.
-        σ: The next batch of samples.
-    """
-    return sampler.sample_next(machine, parameters, state)
 
 
 def MetropolisLocal(hilbert, **kwargs) -> MetropolisSampler: